backend/requirements.txt
# Core Dependencies
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
gitpython==3.1.40
aiohttp==3.9.1
tqdm==4.66.1
lxml==4.9.3 

# FastAPI Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# CORS for frontend communication
fastapi-cors==0.0.6 